Add the option of compile the project inside a docker
With the pourpose of can develop the project in any platform
it has been added the script "build_docker.sh" that compiles
inside a docker container and returns de .img. Finally the user
can execute TRES inside qemu.

- Add a Dockerfile.build that only copys the code
- Add a Dockerfile.environment that builds the environment
  this is a ubuntu with a cross compiler

Is not very efficient and you could use just the binary and not
have to build all the cross compiler but by this way is more
educational and fun.